[VOID] Media View
Added Media Models

Added List and Thumbnail delegates

Media View supports thumnail view and switching between them
