+++ resources/resources.qrc
@@ -13,6 +13,8 @@
         <file>../icons/icon_delete.svg</file>
         <file>../icons/icon_eraser.svg</file>
         <file>../icons/icon_fullscreen.svg</file>
+        <file>../icons/icon_grid_view.svg</file>
+        <file>../icons/icon_list_view.svg</file>
         <file>../icons/icon_maximize.svg</file>
         <file>../icons/icon_minimize.svg</file>
         <file>../icons/icon_missing_frame.svg</file>
@@ -20,6 +22,7 @@
         <file>../icons/icon_repeat.svg</file>
         <file>../icons/icon_restore.svg</file>
         <file>../icons/icon_settings.svg</file>
+        <file>../icons/icon_sort_abc.svg</file>
         <file>../icons/icon_text.svg</file>
         <file>../icons/icon_video_player.png</file>
         <file>../icons/icon_zoom_in.svg</file>
