+++ src/VoidUi/VoidStyle.h
@@ -20,6 +20,7 @@ const QColor VOID_BLUE_COLOR = {30, 210, 200};
 
 const QColor VOID_PURPLE_COLOR = {130, 110, 190};
 
+const QColor VOID_DARK_HIGHLIGHT_GREEN = {80, 110, 30};
 const QColor VOID_HIGHLIGHT_GREEN = {160, 190, 60};
 const QColor VOID_FOREGROUND_COLOR = {190, 190, 190};
